% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae-ratio.R
\name{aer_median_plot}
\alias{aer_median_plot}
\title{Median of actual cumulative loss ratio}
\usage{
aer_median_plot(
  x,
  group_var,
  elapsed_var = "elpm",
  color_type = c("base", "deep"),
  scales = c("fixed", "free_y", "free_x", "free"),
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{x}{an aer.data object}

\item{group_var}{a name of the group variable}

\item{elapsed_var}{a name of the elapsed variable ("elpm", "elp")}

\item{color_type}{a string of color type, base and deep}

\item{scales}{Should \code{scales} be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{theme}{a string specifying a \code{\link[ggshort:switch_theme]{ggshort::switch_theme()}} function ("view", "save", "shiny")}

\item{...}{\code{\link[=theme_view]{theme_view()}}, \code{\link[=theme_save]{theme_save()}}, \code{\link[=theme_shiny]{theme_shiny()}} arguments}
}
\value{
a ggplot object
}
\description{
Draw a median of actual cumulative loss ratio
}
