% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr_comp_plot}
\alias{alr_comp_plot}
\title{Compare actual cumulative loss ratio plot}
\usage{
alr_comp_plot(
  df,
  elapsed_num,
  period_var = "uym",
  elapsed_var = "elpm",
  theme = c("view", "save", "shiny")
)
}
\arguments{
\item{df}{a data.frame}

\item{elapsed_num}{a numeric specifying elapsed months}

\item{period_var}{a name of the period variable ("uym", "uy")}

\item{elapsed_var}{a name of the elapsed variable ("elpm", "elp")}

\item{theme}{a string specifying a \code{\link[ggshort:switch_theme]{ggshort::switch_theme()}} function ("view", "save", "shiny")}
}
\value{
a gtable object
}
\description{
Compare actual comulative loss, risk premium and loss ratio among underwriting year months.
}
