% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\name{amt_plot}
\alias{amt_plot}
\title{Face amount mix plot}
\usage{
amt_plot(amt_mix, label_size = 4, theme = c("view", "save", "shiny"))
}
\arguments{
\item{amt_mix}{a data.frame specifying a face amount mix by gender, age_band}

\item{label_size}{a string specifying label size}

\item{theme}{a string specifying a \code{\link[=match_theme]{match_theme()}} function ("view", "save", "shiny")}
}
\value{
a gtable object
}
\description{
Draw a face amount mix plot.
}
