% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auw-package.R
\docType{package}
\name{auw-package}
\alias{auw}
\alias{auw-package}
\title{auw: Actuarial Underwriting}
\description{
This package is developed for actuarial underwriting.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
\keyword{internal}
