% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{check_cvd_level}
\alias{check_cvd_level}
\title{Check covered surgery levels}
\usage{
check_cvd_level(cvd_level, level)
}
\arguments{
\item{cvd_level}{Integer (or character) vector of covered surgery levels.}

\item{level}{Integer (or character) vector of surgery levels to be checked.}
}
\value{
A binary matrix with \code{length(level)} rows and \code{length(cvd_level)} columns,
where element \verb{[i, j]} is \code{1} if \code{level[i]} matches \code{cvd_level[j]}, otherwise \code{0}.
}
\description{
Create a binary matrix indicating which surgery levels in \code{level}
are covered by each element of \code{cvd_level}. Each column corresponds to one
covered level, and each row corresponds to one element of \code{level}.
}
\examples{
head(check_cvd_level(c(1, 2, 3, 4, 5), c(1, 1, 3, 2, 4, 5, 2, 3, 2, 5)), 20)

}
