% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rider.R
\name{classify_rider}
\alias{classify_rider}
\title{Classify riders}
\usage{
classify_rider(df, rider)
}
\arguments{
\item{df}{A dataframe containing rider information.}

\item{rider}{A string or symbol representing the column name to be categorized.}
}
\value{
The input dataframe with two new columns: \code{coverage_type} and \code{risk_type}.
}
\description{
This function assigns a coverage type (\code{coverage_type}) and a risk type (\code{risk_type})
to a given rider column in a dataframe based on keyword matching.
}
