% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{count_period}
\alias{count_period}
\title{Difference in periods between start and end dates}
\usage{
count_period(start, end, group = 1)
}
\arguments{
\item{start}{A \code{Date} vector. Start date(s).}

\item{end}{A \code{Date} vector. End date(s).}

\item{group}{Integer (>= 1). Group size in months. Default is 1 (monthly).}
}
\value{
An \strong{integer} vector giving the number of periods between \code{start} and
\code{end}, grouped by \code{group} months. \code{NA} dates yield \code{NA} results.
}
\description{
Compute the number of grouped periods (in months) between two dates.
}
\examples{
count_period(as.Date("2020-01-01"), as.Date("2021-01-01"), group = 12)
count_period(as.Date(c("2020-01-15","2020-02-01")),
             as.Date(c("2020-03-01","2020-02-20")), group = 1)

}
