% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{get_diz_list}
\alias{get_diz_list}
\title{Get reserved disease list}
\usage{
get_diz_list()
}
\value{
A named list of reserved-disease objects found in \code{.AUW_ENV}.
}
\description{
Return all reserved disease objects stored in the internal package environment.
}
\examples{
\donttest{
diz_list <- get_diz_list()
names(diz_list)
}

}
\seealso{
\code{\link[=get_kcd]{get_kcd()}}, \code{\link[=add_kcd_name]{add_kcd_name()}}
}
