% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inforce.R
\name{get_inforce_ym_period}
\alias{get_inforce_ym_period}
\title{Inforce statistic by year, month, period}
\usage{
get_inforce_ym_period(df, id_var, group_var, from_var, to_var)
}
\arguments{
\item{df}{a data.frame}

\item{id_var}{a name of id variable}

\item{group_var}{names of group variables}

\item{from_var}{a name of start date variable}

\item{to_var}{a name of end date variable}
}
\value{
a data frame
}
\description{
Get inforce statistic by year, month, period
}
