% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{get_rr}
\alias{get_rr}
\title{Get relative risks}
\usage{
get_rr(x, decl_vs = c("0", "1"), threshold = 0.975)
}
\arguments{
\item{x}{a rr object}

\item{decl_vs}{a two elements string vector contains declaration items to compare.}

\item{threshold}{threshold to reject null hypothesis. (default: 0.975)}
}
\value{
a data frame contains relative risks
}
\description{
Get relative risks from the data.rr object.
}
