% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{get_stat_alr}
\alias{get_stat_alr}
\title{Get stat from actual loss ratio}
\usage{
get_stat_alr(
  df,
  group_var,
  value_var = c("loss", "rp"),
  period_var = c("uym"),
  elapsed_var = c("elpm")
)
}
\arguments{
\item{df}{a data.frame contains actual loss ratio}

\item{group_var}{name of columns}

\item{value_var}{name of columns specifying loss, rp}

\item{period_var}{a name of column specifying period}

\item{elapsed_var}{a name of column specifying elapsed period}
}
\value{
a data.frame
}
\description{
Get stat from actual loss ratio data (To-be updated)
}
