% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{id_with_kcd_terms}
\alias{id_with_kcd_terms}
\title{id with kcd terms}
\usage{
id_with_kcd_terms(df, id_group_var, kcd_var, from_var, to_var, udate, ...)
}
\arguments{
\item{df}{a cohort data frame}

\item{id_group_var}{names of id and group variables like list(id, group, age_band).
If no group variables, only a name of id variable.}

\item{kcd_var}{a name of kcd variable}

\item{from_var}{a name of start date variable}

\item{to_var}{a name of end date variable}

\item{udate}{a string date underwritten}

\item{...}{the basic form of kcd terms is like "decl = list(start, end, kcd codes)",
start means xx months before udate (negative numeric) and end means xx months
after udate (positive months) (ex. list(-36, 36, "M51"))}
}
\value{
a data frame
}
\description{
Whether the ids have KCD codes that exists before or after a specific point
in time
}
\details{
Subset cohort data using underwriting date and kcd code
}
\examples{
# subset cohort
\dontrun{
id_with_kcd_terms(cohort, id, list(age_band), kcd, sdate, edate, udate,
                  list(-60,  0, "I10"),
                  list(-60,  0, "I2[0-5]|I6[0-9]|G46"),
                  list(  0, 36, "I2[0-5]|I6[0-9]|G46"))}

}
