% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{id_with_rr_kcd_terms}
\alias{id_with_rr_kcd_terms}
\title{id with "relative risk" kcd terms}
\usage{
id_with_rr_kcd_terms(
  df,
  id_group_var,
  kcd_var,
  from_var,
  to_var,
  udate,
  decl1 = NULL,
  decl2 = NULL,
  decl3 = NULL,
  excl = NULL,
  claim = NULL
)
}
\arguments{
\item{df}{a cohort data frame}

\item{id_group_var}{names of id and group variables like .(id, group, age_band)
or list(id, age_band). If no group variables, only a name of id variable.}

\item{kcd_var}{a name of kcd variable}

\item{from_var}{a name of start date variable}

\item{to_var}{a name of end date variable}

\item{udate}{a string date underwritten}

\item{decl1, decl2, decl3, excl, claim}{the basic form of kcd terms is like
"decl1 = list(start, end, kcd codes)", start means xx months before udate
(negative numeric) and end means xx months after udate (positive months)
(ex. decl1 = list(-36, 36, "M51")). decl means declaration, excl means exclusion,
claim means claim.}
}
\value{
a data frame
}
\description{
Whether the ids have KCD codes that exists before or after a specific point
in time
}
\details{
Subset cohort data using underwriting date and kcd code
}
\examples{
# id with kcd terms for 'Relative Risk'
\dontrun{
id_with_rr_kcd_terms(
  df = cohort,
  id_group_var = .(id, gender, age_band),
  kcd_var = kcd,
  from_var = sdate,
  to_var = edate,
  udate = udate,
  decl1 = list(-60,  0, "I10"),
  decl2 = list(-60,  0, "E78"),
  decl3 = list(-60,  0, "E1[0-4]"),
  excl  = list(-60,  0, "I2[0-5]|I6[0-9]|G46"),
  claim = list(  0, 36, "I2[0-5]|I6[0-9]|G46")
)}

}
