% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-risk.R
\name{irplot}
\alias{irplot}
\title{Deprecated: irplot()}
\usage{
irplot(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=ir_plot]{ir_plot()}}.}
}
\value{
Same return value as \code{\link[=ir_plot]{ir_plot()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=ir_plot]{ir_plot()}} instead.
}
\seealso{
\code{\link[=ir_plot]{ir_plot()}}
}
