% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kcd.R
\docType{data}
\name{kcd_book}
\alias{kcd_book}
\title{KCD data}
\format{
a data.table with two four variables and about twenty-two thousands observations
}
\usage{
kcd_book
}
\description{
This data set describes KCD (Korean Standard Classification of Diseases) codes.
}
\examples{
kcd_book
}
\keyword{datasets}
