% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae-ratio.R
\name{plot_aer}
\alias{plot_aer}
\title{Actual A/E ratio by each UY months}
\usage{
plot_aer(
  x,
  group_var,
  period_var = "uym",
  elapsed_var = "elpm",
  value_var = "caer",
  scales = c("fixed", "free_y", "free_x", "free"),
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{x}{an aer object}

\item{group_var}{a name of the group variable}

\item{period_var}{a name of the period variable ("uym", "uy")}

\item{elapsed_var}{a name of the elapsed variable ("elpm", "elp")}

\item{scales}{Should \code{scales} be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{theme}{a string specifying a \code{\link[ggshort:switch_theme]{ggshort::switch_theme()}} function ("view", "save", "shiny")}

\item{...}{\code{\link[=theme_view]{theme_view()}}, \code{\link[=theme_save]{theme_save()}}, \code{\link[=theme_shiny]{theme_shiny()}} arguments}

\item{index_var}{a name of the index variable ("caer", "closs", "crp")}
}
\value{
a ggplot object
}
\description{
Draw an actual loss ratio by each UY months.
}
