% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-risk.R
\name{rrplot}
\alias{rrplot}
\title{Deprecated: rrplot()}
\usage{
rrplot(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=rr_plot]{rr_plot()}}.}
}
\value{
Same return value as \code{\link[=rr_plot]{rr_plot()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=rr_plot]{rr_plot()}} instead.
}
\seealso{
\code{\link[=rr_plot]{rr_plot()}}
}
