% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{set_age_band}
\alias{set_age_band}
\title{Set age band}
\usage{
set_age_band(
  df,
  age_var,
  interval = 5,
  right = FALSE,
  col_nm = "age_band",
  labels
)
}
\arguments{
\item{df}{a data.frame}

\item{age_var}{a name of column specifying age column}

\item{interval}{a numeric specifying an age inteval (default: 5)}

\item{right}{logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}

\item{col_nm}{a string sepcifying age band column name}

\item{labels}{a string vector specifying age band label}
}
\value{
no return value
}
\description{
Set age band
}
\examples{
\donttest{df <- data.frame(age = sample(0:99, 20, replace = TRUE))
set_age_band(df, age)}

}
