% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{set_band}
\alias{set_band}
\title{Set bands}
\usage{
set_band(
  df,
  var,
  breaks,
  interval = 5,
  right = FALSE,
  col_nm,
  cutoff = FALSE,
  label_type = c("close", "open.start", "open", "open.end")
)
}
\arguments{
\item{df}{a data.frame}

\item{var}{a name of column}

\item{breaks}{either a numeric vector of two or more unique cut points or a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut.}

\item{interval}{a numeric specifying interval (default: 5)}

\item{right}{logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}

\item{col_nm}{a string specifying a column name}

\item{cutoff}{a logical to use maximum age cutoff}

\item{label_type}{a string vector specifying label type ("close", "open", "open.start", "open.end")}
}
\value{
no return value
}
\description{
Set bands
}
\examples{
\dontrun{
df <- mtcars
set_band(df, age)}

}
