% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{set_kcd_name}
\alias{set_kcd_name}
\title{Deprecated: set_kcd_name()}
\usage{
set_kcd_name(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=add_kcd_name]{add_kcd_name()}}.}
}
\value{
Same return value as \code{\link[=add_kcd_name]{add_kcd_name()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=add_kcd_name]{add_kcd_name()}} instead.
}
\seealso{
\code{\link[=add_kcd_name]{add_kcd_name()}}
}
