% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{set_kcd_sub}
\alias{set_kcd_sub}
\title{Set kcd sub columns}
\usage{
set_kcd_sub(df, kcd_var, digit = c(3L, 2L, 1L))
}
\arguments{
\item{df}{a data.frame}

\item{kcd_var}{a name of a kcd column}

\item{digit}{a stop digit of a string vector}
}
\value{
no return values.
}
\description{
Set kcd sub columns.
}
\examples{
# Set kcd sub columns
\dontrun{
set_kcd_sub(df, kcd)}

}
