% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{split_date}
\alias{split_date}
\title{Split date range by specific date}
\usage{
split_date(df, from_var, to_var, udate, all = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{a cohort data frame}

\item{from_var}{a name of start date variable}

\item{to_var}{a name of end date variable}

\item{udate}{a underwriting date}

\item{all}{a boolean specifying only for splited data or}

\item{verbose}{a boolean giving verbose information}
}
\value{
a data frame
}
\description{
Split date range by specific date.
}
