% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_id_with_kcd}
\alias{subset_id_with_kcd}
\title{Subset cohort data}
\usage{
subset_id_with_kcd(
  df,
  id_var,
  kcd_var,
  from_var,
  to_var,
  udate,
  start,
  end,
  ...
)
}
\arguments{
\item{df}{a cohort data frame}

\item{id_var}{a name of id variable}

\item{kcd_var}{a name of kcd variable}

\item{from_var}{a name of start date variable}

\item{to_var}{a name of end date variable}

\item{udate}{a string date underwritten}

\item{start}{xx months before udate (negative numeric)}

\item{end}{xx months after udate (positive numeric)}

\item{...}{kcd codes}
}
\description{
Subset cohort data using underwriting date and kcd code
}
\examples{
# subset cohort
\dontrun{
subset_id_with_kcd(cohort, id, kcd, sdate, edate, udate, -60, 36, "M51")}

}
