% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-risk.R
\name{summary.id_data}
\alias{summary.id_data}
\title{Summarise an \code{id_data} object}
\usage{
\method{summary}{id_data}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"id_data"}.}

\item{...}{Unused.}
}
\value{
A restored summary table stored as an attribute of \code{object}:
\code{attr(object, "summary")}
}
\description{
Summarise an \code{id_data} object
}
