% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr_uym_plot}
\alias{alr_uym_plot}
\title{Actual loss ratio by each UY months}
\usage{
alr_uym_plot(
  x,
  group_var,
  period_var = "uym",
  elapsed_var = "elpm",
  index_var = "clr",
  scales = c("fixed", "free_y", "free_x", "free"),
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{x}{an alr.data object}

\item{group_var}{a name of the group variable}

\item{period_var}{a name of the period variable ("uym", "uy")}

\item{elapsed_var}{a name of the elapsed variable ("elpm", "elp")}

\item{index_var}{a name of the index variable ("clr", "closs", "crp")}

\item{scales}{Should \code{scales} be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{theme}{a string specifying a \code{\link[=match_theme]{match_theme()}} function ("view", "save", "shiny")}

\item{...}{\code{\link[=theme_view]{theme_view()}}, \code{\link[=theme_save]{theme_save()}}, \code{\link[=theme_shiny]{theme_shiny()}} arguments}
}
\value{
a ggplot object
}
\description{
Draw an actual loss ratio by each UY months.
}
