% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\name{amt_plot}
\alias{amt_plot}
\title{Face amount mix plot}
\usage{
amt_plot(
  amt_mix,
  label_args = list(size = 4),
  theme = c("view", "save", "shiny")
)
}
\arguments{
\item{amt_mix}{a data.frame specifying a face amount mix by gender, age_band}

\item{theme}{a string specifying a \code{\link[ggshort:switch_theme]{ggshort::switch_theme()}} function ("view", "save", "shiny")}

\item{label_size}{a string specifying label size}
}
\value{
a gtable object
}
\description{
Draw a face amount mix plot.
}
