% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{comp_risk_plot}
\alias{comp_risk_plot}
\title{Compare two risk rates}
\usage{
comp_risk_plot(
  risk_info,
  risk1,
  risk2,
  plot = TRUE,
  logscale = FALSE,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  age_unit = 10,
  widths = c(5, 3, 2),
  heights = c(2, 8),
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{risk_info}{a data frame specifying risk_rate}

\item{risk1}{a string specifying risk 1}

\item{risk2}{a string specifying risk 2}

\item{plot}{a logical whether to show a plot or not}

\item{logscale}{a logical whether to use a log scale}

\item{nrow, ncol}{Number of rows and columns.}

\item{scales}{Should scales be fixed ("\code{fixed}", the default), free ("\code{free}"),
or free in one dimension ("\code{free_x}","\code{free_y}")?}

\item{age_unit}{age interval unit to express}

\item{widths}{a unit vector giving the height of each row or the widths of 3 columns
(risk rate plot, ratio plot, legend)}

\item{heights}{a unit vector giving the height of 2 row (top and plot)}

\item{theme}{a string specifying a ggshort theme function ("view", "save", "shiny")}

\item{...}{arguments passed on to (theme_view, theme_save, theme_shiny)}
}
\value{
no return values
}
\description{
Compare two risk rates.
}
\examples{
# compare two risk rates
\dontrun{
comp_risk_plot(risk_info, risk1, risk2)}

}
