% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{get_kcd}
\alias{get_kcd}
\title{Get KCD codes}
\usage{
get_kcd(x, lang = c("ko", "en"), type = c("kcd", "ko", "en"))
}
\arguments{
\item{x}{a KCD code string or regular expression.}

\item{lang}{a string specifying language to explain ("ko", "en")}

\item{type}{a string sepcifying a type of a return value ("kcd", "ko", "en")}
}
\value{
a string vector
}
\description{
Search for KCD codes in Korean or English
}
\examples{
# get a kcd code
\donttest{get_kcd("M51")
get_kcd("M51$|S33$")
get_kcd("Diabetes")}

# get a related kcd code vector
\donttest{kcd <- get_kcd("M51$|S33$", type = "kcd")
kcd}

}
