% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{get_search_pattern}
\alias{get_search_pattern}
\title{Get pattern to search strings}
\usage{
get_search_pattern(x, collapse = "|")
}
\arguments{
\item{x}{a string vector}

\item{collapse}{an optional character string to separate the results. Not
\code{\link{NA_character_}}. When \code{collapse} is a string, the result is always a string
(\code{\link{character}} of length 1). default "|"}
}
\value{
a string vector
}
\description{
Get regular expression pattern to search strings
}
\examples{
# get regular expression to search string
\donttest{get_search_pattern(c("a|b", "a|c", "c|de|f"))}

}
