% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{irplot}
\alias{irplot}
\title{Incidence Rate Plot}
\usage{
irplot(
  x,
  color_type = c("base", "deep"),
  scales = c("fixed", "free_y", "free_x", "free"),
  theme = c("view", "save", "shiny")
)
}
\arguments{
\item{x}{ir object}

\item{color_type}{a string specifying a pair_color_type ("base", "deep")}

\item{scales}{Should scales be fixed ("fixed", the default), free ("free"),
or free in one dimension ("free_x", "free_y")?}

\item{theme}{a string specifying a \code{\link[=match_theme]{match_theme()}} function ("view", "save", "shiny")}
}
\value{
a ggplot object
}
\description{
Draw an incidence rate plot.
}
\examples{
# draw an incidence rate plot
\dontrun{
irplot(x)}

}
