% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae-ratio.R, R/alr.R
\name{longer}
\alias{longer}
\title{Longer data structure}
\usage{
longer(data, ...)

longer(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Make a data structure longer

Make a data structure longer
}
