% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_plot}
\alias{risk_plot}
\title{Risk rate plot}
\usage{
risk_plot(
  risk_info,
  x,
  logscale = FALSE,
  max_label = TRUE,
  label_family = "Comic Sans MS",
  age_interval = 10,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{risk_info}{a data frame specifying risk_rate}

\item{x}{a string vector specifying risk}

\item{logscale}{a boolean specifying a log scale}

\item{max_label}{a boolean whether to draw a max label or not}

\item{label_family}{a string specifying label font-family}

\item{age_interval}{a numeric specifying an age interval for the x label (default: 10)}

\item{nrow, ncol}{Number of rows and columns.}

\item{scales}{Should \code{scales} be fixed ("\code{fixed}", the default), free ("\code{free}"), or free in one
dimension ("\code{free_x}", "\code{free_y}")?}

\item{theme}{a string specifying a \code{\link[=match_theme]{match_theme()}} function ("view", "save", "shiny")}

\item{...}{arguments passed on to (theme_view, theme_save, theme_shiny)}
}
\value{
a ggplot object
}
\description{
Draw a risk rate plot.
}
\examples{
# draw a risk rate plot
\dontrun{risk_plot(risk_info)}

}
