% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{rrplot}
\alias{rrplot}
\title{Relative Risk Plot}
\usage{
rrplot(
  x,
  logscale = FALSE,
  scales = c("fixed", "free_y", "free_x", "free"),
  theme = c("view", "save", "shiny")
)
}
\arguments{
\item{x}{a rr object}

\item{logscale}{a boolean specifying a log scale}

\item{scales}{Should \code{scales} be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{theme}{a string specifying a \code{\link[=match_theme]{match_theme()}} function ("view", "save", "shiny")}
}
\value{
a ggplot object
}
\description{
Draw a relative risk plot
}
\examples{
# relative risk plot
\dontrun{
rrplot(x)}

}
