% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{set_age_band}
\alias{set_age_band}
\title{Deprecated: set_age_band()}
\usage{
set_age_band(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=add_age_band]{add_age_band()}}.}
}
\value{
Same return value as \code{\link[=add_age_band]{add_age_band()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=add_age_band]{add_age_band()}} instead.
}
\seealso{
\code{\link[=add_age_band]{add_age_band()}}
}
