% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{set_kcd_name}
\alias{set_kcd_name}
\title{Set kcd name}
\usage{
set_kcd_name(df, kcd_var, dots = TRUE, lang = c("ko", "en"))
}
\arguments{
\item{df}{a data.frame}

\item{kcd_var}{a name of a kcd column}

\item{dots}{a logical whether a kcd column contains dots}

\item{lang}{a string specifying language ("ko", "en")}
}
\value{
no return values.
}
\description{
Set kcd name (en, ko).
}
\examples{
# Set kcd sub columns
\dontrun{
set_kcd_sub(df, kcd)}

}
