% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcd.R
\name{set_kcd_sub}
\alias{set_kcd_sub}
\title{Deprecated: set_kcd_sub()}
\usage{
set_kcd_sub(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=add_kcd_sub]{add_kcd_sub()}}.}
}
\value{
Same return value as \code{\link[=add_kcd_sub]{add_kcd_sub()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=add_kcd_sub]{add_kcd_sub()}} instead.
}
\seealso{
\code{\link[=add_kcd_sub]{add_kcd_sub()}}
}
